Move the allocation related functions to a separate class to avoid code blow, as well as add separate test for it.
