+++ lib/money.rb
@@ -2,6 +2,7 @@ require_relative 'money/money_parser'
 require_relative 'money/helpers'
 require_relative 'money/currency'
 require_relative 'money/null_currency'
+require_relative 'money/allocator'
 require_relative 'money/money'
 require_relative 'money/errors'
 require_relative 'money/deprecations'
