+++ lib/money/money.rb
@@ -245,114 +245,14 @@ class Money
     Money.new(result, currency)
   end
 
-  # Allocates money between different parties without losing pennies.
-  # After the mathematically split has been performed, left over pennies will
-  # be distributed round-robin amongst the parties. This means that parties
-  # listed first will likely receive more pennies than ones that are listed later
-  #
-  # @param splits [Array<Numeric>]
-  # @param strategy Symbol
-  # @return [Array<Money>]
-  #
-  # Strategies:
-  # - `:roundrobin` (default): leftover pennies will be accumulated starting from the first allocation left to right
-  # - `:roundrobin_reverse`: leftover pennies will be accumulated starting from the last allocation right to left
-  #
-  # @example
-  #   Money.new(5, "USD").allocate([0.50, 0.25, 0.25])
-  #     #=> [#<Money value:2.50 currency:USD>, #<Money value:1.25 currency:USD>, #<Money value:1.25 currency:USD>]
-  #   Money.new(5, "USD").allocate([0.3, 0.7])
-  #     #=> [#<Money value:1.50 currency:USD>, #<Money value:3.50 currency:USD>]
-  #   Money.new(100, "USD").allocate([0.33, 0.33, 0.33])
-  #     #=> [#<Money value:33.34 currency:USD>, #<Money value:33.33 currency:USD>, #<Money value:33.33 currency:USD>]
-
-  # @example left over cents distributed to first party due to rounding, and two solutions for a more natural distribution
-  #   Money.new(30, "USD").allocate([0.667, 0.333])
-  #     #=> [#<Money value:20.01 currency:USD>, #<Money value:9.99 currency:USD>]
-  #   Money.new(30, "USD").allocate([0.333, 0.667])
-  #     #=> [#<Money value:20.00 currency:USD>, #<Money value:10.00 currency:USD>]
-  #   Money.new(30, "USD").allocate([Rational(2, 3), Rational(1, 3)])
-  #     #=> [#<Money value:20.00 currency:USD>, #<Money value:10.00 currency:USD>]
-
-  # @example left over pennies distributed reverse order when using roundrobin_reverse strategy
-  #   Money.new(10.01, "USD").allocate([0.5, 0.5], :roundrobin_reverse)
-  #     #=> [#<Money value:5.00 currency:USD>, #<Money value:5.01 currency:USD>]
+  # @see Money::Allocator#allocate
   def allocate(splits, strategy = :roundrobin)
-    if all_rational?(splits)
-      allocations = splits.inject(0) { |sum, n| sum + n }
-    else
-      allocations = splits.inject(0) { |sum, n| sum + Helpers.value_to_decimal(n) }
-    end
-
-    if (allocations - BigDecimal("1")) > Float::EPSILON
-      raise ArgumentError, "splits add to more than 100%"
-    end
-
-    amounts, left_over = amounts_from_splits(allocations, splits)
-
-    left_over.to_i.times do |i|
-      idx = case strategy
-      when :roundrobin
-        i % amounts.length
-      when :roundrobin_reverse
-        amounts.length - (i % amounts.length) - 1
-      else
-        raise ArgumentError, "Invalid strategy. Valid options: :roundrobin, :roundrobin_reverse"
-      end
-
-      amounts[idx] += 1
-    end
-
-    amounts.collect { |subunits| Money.from_subunits(subunits, currency) }
+    Money::Allocator.new(self).allocate(splits, strategy)
   end
 
-  # Allocates money between different parties up to the maximum amounts specified.
-  # Left over pennies will be assigned round-robin up to the maximum specified.
-  # Pennies are dropped when the maximums are attained.
-  #
-  # @example
-  #   Money.new(30.75).allocate_max_amounts([Money.new(26), Money.new(4.75)])
-  #     #=> [Money.new(26), Money.new(4.75)]
-  #
-  #   Money.new(30.75).allocate_max_amounts([Money.new(26), Money.new(4.74)]
-  #     #=> [Money.new(26), Money.new(4.74)]
-  #
-  #   Money.new(30).allocate_max_amounts([Money.new(15), Money.new(15)]
-  #     #=> [Money.new(15), Money.new(15)]
-  #
-  #   Money.new(1).allocate_max_amounts([Money.new(33), Money.new(33), Money.new(33)])
-  #     #=> [Money.new(0.34), Money.new(0.33), Money.new(0.33)]
-  #
-  #   Money.new(100).allocate_max_amounts([Money.new(5), Money.new(2)])
-  #     #=> [Money.new(5), Money.new(2)]
+  # @see Money::Allocator#allocate_max_amounts
   def allocate_max_amounts(maximums)
-    allocation_currency = extract_currency(maximums + [self])
-    maximums = maximums.map { |max| max.to_money(allocation_currency) }
-    maximums_total = maximums.reduce(Money.new(0, allocation_currency), :+)
-
-    splits = maximums.map do |max_amount|
-      next(0) if maximums_total.zero?
-      Money.rational(max_amount, maximums_total)
-    end
-
-    total_allocatable = [
-      value * allocation_currency.subunit_to_unit,
-      maximums_total.value * allocation_currency.subunit_to_unit
-    ].min
-
-    subunits_amounts, left_over = amounts_from_splits(1, splits, total_allocatable)
-
-    subunits_amounts.each_with_index do |amount, index|
-      break unless left_over > 0
-
-      max_amount = maximums[index].value * allocation_currency.subunit_to_unit
-      next unless amount < max_amount
-
-      left_over -= 1
-      subunits_amounts[index] += 1
-    end
-
-    subunits_amounts.map { |cents| Money.from_subunits(cents, allocation_currency) }
+    Money::Allocator.new(self).allocate_max_amounts(maximums)
   end
 
   # Split money amongst parties evenly without losing pennies.
@@ -415,22 +315,6 @@ class Money
 
   private
 
-  def all_rational?(splits)
-    splits.all? { |split| split.is_a?(Rational) }
-  end
-
-  def amounts_from_splits(allocations, splits, subunits_to_split = subunits)
-    left_over = subunits_to_split
-
-    amounts = splits.collect do |ratio|
-      frac = (Helpers.value_to_decimal(subunits_to_split * ratio) / allocations).floor
-      left_over -= frac
-      frac
-    end
-
-    [amounts, left_over]
-  end
-
   def arithmetic(money_or_numeric)
     raise TypeError, "#{money_or_numeric.class.name} can't be coerced into Money" unless money_or_numeric.respond_to?(:to_money)
     other = money_or_numeric.to_money(currency)
@@ -440,13 +324,5 @@ class Money
 
   def calculated_currency(other)
     no_currency? ? other : currency
-  end
-
-  def extract_currency(money_array)
-    currencies = money_array.lazy.select { |money| money.is_a?(Money) }.reject(&:no_currency?).map(&:currency).to_a.uniq
-    if currencies.size > 1
-      raise ArgumentError, "operation not permitted for Money objects with different currencies #{currencies.join(', ')}"
-    end
-    currencies.first || NULL_CURRENCY
   end
 end
